Test object serialization and deserialization
