+++ api/src/test/java/cz/vse/java/hangman/api/messages/DummyMessage.java
@@ -1,4 +1,4 @@
-package cz.vse.java.hangman.api;
+package cz.vse.java.hangman.api.messages;
 
 public record DummyMessage(int id, String name, String desc) implements Message {
 }
