+++ api/src/test/java/cz/vse/java/hangman/api/messages/DummyMessageWithObject.java
@@ -1,0 +1,4 @@
+package cz.vse.java.hangman.api.messages;
+
+public record DummyMessageWithObject(DummyObject dummyObject) implements Message {
+}
