+++ api/src/test/java/cz/vse/java/hangman/api/messages/DummyObject.java
@@ -1,0 +1,4 @@
+package cz.vse.java.hangman.api.messages;
+
+public record DummyObject(String param) {
+}
